Now mindmap working without refresh, delete mimdmap.js and insert data in TreeWrapper.js some modify in chat.js related to mindmap
