+++ package.json
@@ -55,6 +55,7 @@
     "passport": "^0.4.1",
     "pg": "^7.14.0",
     "react-bootstrap": "^1.0.0-beta.16",
+    "react-faux-dom": "^4.5.0",
     "react-router": "^5.1.2",
     "sequelize": "^5.21.2",
     "socket.io": "^2.3.0",
